% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{ATE}
\alias{ATE}
\title{ATE of a matched dataset}
\usage{
ATE(AHB_out)
}
\arguments{
\item{AHB_out}{An object returned by running \code{\link{AHB_fast_match}} and
\code{\link{AHB_MIP_match}}}
}
\description{
\code{ATE} computes the average treatment effect (ATE) of a matched dataset.
}
\details{
The ATE is computed as the difference between the weighted treated and the
weighted control outcomes in the dataset. A unit's weight is the number of
times it was matched.
}
