% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate Toy Data for AHB Matching}
\usage{
gen_data(n_units = 100, p = 5)
}
\arguments{
\item{n_units}{Number of units desired in the data set. Defaults to 100.}

\item{p}{Number of covariates in the data set. Must be greater than 2.
Defaults to 5.}
}
\value{
A data frame that may be passed to  \code{\link{AHB_fast_match}} and
  \code{\link{AHB_MIP_match}}. Covariates are continuous. Treatment is binary
  numeric and outcome is numeric.
}
\description{
\code{gen_data} generates toy data that can be used to explore AHB's
functionality.
}
\details{
\code{gen_data} simulates data in the format accepted by
\code{\link{AHB_fast_match}} and \code{\link{AHB_MIP_match}}. Covariates
\eqn{X} and treatment \eqn{T} are independently generated according to
uniformly distribution and Bernoulli(0.5) distribution respectively. The
outcome \eqn{Y} is generated according to \eqn{Y = 2 + 5*(X[,1] > 1.5)*T +
\epsilon}, where \eqn{\epsilon \sim N(0, I_n)}. Thus, the value of \code{p}
must be at least 2.
}
