% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{print_box}
\alias{print_box}
\title{\code{print_box}  prints the box info on original covariates
 constructed based on a specific treated unit.}
\usage{
print_box(AHB_out, treated_unit_id, returnValue = F)
}
\arguments{
\item{AHB_out}{An object returned by running \code{\link{AHB_fast_match}} and
\code{\link{AHB_MIP_match}}}

\item{treated_unit_id}{is the treated unit id}

\item{returnValue}{is if we return the info or not}
}
\value{
a list that contains the box info
}
\description{
\code{print_box}  prints the box info on original covariates
 constructed based on a specific treated unit.
}
